rename mock_backend to mock_amdsmi_backend
