+++ projects/rocprofiler-systems/source/lib/backends/amd_smi/tests/mock_amdsmi_backend.hpp
