+++ projects/rocprofiler-systems/source/lib/backends/amd_smi/tests/test_backend.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "backends/amd_smi/backend.hpp"
-#include "mock_backend.hpp"
+#include "mock_amdsmi_backend.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
