+++ projects/rocprofiler-systems/source/lib/backends/amd_smi/tests/test_device_backend.cpp
@@ -3,7 +3,7 @@
 
 #include "backends/amd_smi/backend.hpp"
 #include "backends/amd_smi/device_backend.hpp"
-#include "mock_backend.hpp"
+#include "mock_amdsmi_backend.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
