+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/device_providers/amd_smi/tests/test_provider.cpp
@@ -4,7 +4,7 @@
 #include "backends/amd_smi/backend.hpp"
 #include "backends/amd_smi/device_backend.hpp"
 #include "library/pmc/device_providers/amd_smi/provider.hpp"
-#include "mock_backend.hpp"
+#include "mock_amdsmi_backend.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
